microservices/REPORT/requirements.txt
confluent-kafka
kafka-python
fpdf2
matplotlib
pandas
hugchat
transformers
torch
sentencepiece
protobuf
python-dotenv
openai
accelerate